Stable ordering when generating StructMeta code
Currently, when user specifies `derive(StructMeta)`, this crate
generates a branch of code for each of the attributes.

These attributes however have so far been stored inside a `HashMap`
which means the ordering of the branches is more or less random. This
has no real impact on the runtime: the order of the branches doesn't
change the semantics of the code. It does matter however for compile
time: `rustc` will calculate a different crate hash (effectively a sort
of ABI signature) if code is re-ordered.

This means that any time we recompile code with `derive(StructMeta)`
somewhere, we're breaking the ABI even if there were _no_ changes at all
to any code. Basically, we're throwing a dice at compile time as to what
order the code will be laid out in.

This is very bad for reproducability and plays very poorly with any sort
of binary caching.

I initially thought that the reproducability issue was coming from
`rustc` itself but the more I was debugging, I made my way back to this
crate.

You can see the original rustc issue for context
[here](https://github.com/rust-lang/rust/issues/89904).

I think this change should be basically invisible to anyone that's been
using the crate so far.